==> analysis/01_TransformData.Rmd <==
---
title: "Pre-processing of FCS files: Arcsih Transformation"
author: "Carolina"
date: "2024-12-16"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This is an R Markdown document to pre-process FCS files by applying transformation to parameter values. 

##Load packages and utilities
```{r, warning=FALSE}

library(tidyverse)
library(flowCore)
library(flowVS)
library(here)

source(here("src/Preprocessing_Functions.R"))

seed <- 1234
set.seed(seed)

```

##Paths to directories and files
*fcsdir* is the path to the directory containing the FCS files. \
*metafile* is the path to the metadata file, which is a csv file containing the sample name (same name as fcs file) in the first column, as well as all the information for each sample (e.g. mouseID, Treatment, Sex, Tissue, etc.). \
*outdir* is the path to the output directory that will contain the output data after transformation. The directory will be created if it doesn't exist.

```{r}
# Set directories
fcsdir <- here("data/FCS/")
metafile <- here("data/metadata.csv")
outdir <- here("data/TransformedData/")

```

## Read data
The `read.fcs` function reads the metadata and identify the fcs files in the fcs directory to read them and create a flowset object. It will also print the fcs file names that are included in the flowset, and will inform of any excluded (present in fcsdir but not in metadata), or missing (in metadata, but not found in fcsdir) fcs file.

```{r}

metadata <- read.csv(metafile)

fset <- read.fcs(fcs.dir.path=fcsdir,
                 metadata=metadata,
                 transform=FALSE,
                 downsample=5000,
                 truncate.max=FALSE)

```

## Obatin parameters to transform
Copy and paste a representative sample from the list above to the variable "representative_sample". \
The following code will create a data frame from the representative sample containing the channels, labels, markers, and an initial cofactor value for transformation. \
Then, it will print a list of the available markers and assigned cofactor, which can be copied and paste to the next section to adjust the cofactor values.

```{r}
# Set the variables. Rerun as needed.
representative_sample <- "<Sample_001.fcs>"
Label_separator <- NULL
initial_cofactor <- 200

###################################################################

# Data frame containing the channels and respective marker names
fftest <- fset@frames[[representative_sample]]

param.df <- process.param (fftest,
                           cofactor=initial_cofactor,
                           label.sep=Label_separator)

```

##Test cofactors
Flow data requires transformation (e.g. log, biexponential,etc). Hyperbolic arcsine (arcsinh) transformation is recommended in this case and it is calculated based on a determined cofactor. \
In the following code the cofactor for each marker can be adjusted then a transformation will be performed. \
Density plots will be printed for each marker. \
The cofactors can be re-adjusted and the code re-run as needed. \
If needed, another representative sample can be tested by editing the previous section.
```{r}

# Adjust cofactors and re-run as needed
custom.cof <- c(
#copy markers and cofactors here as:
'param1'=1000,
'param2'=1000,
#etc.
)

##############################################################################

# Test the cofactors with density plots

test.cofactors(flowframe.obj = fftest,
               param.df = param.df,
               new.cofactors = custom.cof)

```

##Transform the flowset
After choosing the cofactors, the transformation can applied to the entire dataset with the following code.
```{r}

# Transform all the flowset with the redefined cofactors.
fst <- transFlowVS(fset, channels=as.character(param.df$channel), cofactor=param.df$cofactor)

# Obtain a joined expression matrix from the flowfset.
# First column should contain the sample ID.
mat <- expr.fs(fst, param.df)

glimpse(mat)

```

##Save transformed data
The output files will be saved in the pre-established output directory: \
- A csv file with the transformed expression values for the entire dataset. \
- FCS files with the transformed values.
```{r}

# Create directory if it does not exist
if (!file.exists(outdir)){
    dir.create(outdir)
}

# Save flowframes wihtin flowset as fcs files using the flowCore package
write.flowSet(fst, outdir=outdir, filename = sampleNames(fst))

# Save the transformed expression matrix
write.csv(mat, file.path(outdir, "TransfExprData.csv"), row.names=FALSE)

```
