==> analysis/02a_SOM_clustering.Rmd <==
---
title: "FlowSOM clustering"
author: "Carolina"
date: "2025-01-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

There are two main methods for clustering flow cytometry data, FlowSOM and Phenograph. \
Each of them has their own strengths and limitations. \
FlowSOM requires to pre-define the number of clusters that we want to obtain, while in phenograph will give you the number of clusters based on the "k" parameter. However, flowSOM may be preferred for large data sets, as phenograph may return too many clusters.


## Load packages and utilities
```{r}

# Load required packages
library(tidyverse)
library(FlowSOM)
library(umap)
library(Rtsne)
library(here)

source(here("src/Plot_functions_flow.R"))

seed <- 1234
set.seed(seed)
```

##Paths to directories and files
*transfcsv* is the path to the csv file containing the transformed expression matrix. \
*metafile* is the path to the metadata file, which is a csv file containing the sample name (same name as fcs file) in the first column, as well as all the information for each sample (e.g. mouseID, Treatment, Sex, Tissue, etc.). \
*outdir* is the path to the output directory that will contain the output data after clustering. The directory will be created if it doesn't exist.

```{r}

# Set directories
transfcsv <- here("data/TransformedData/TransfExprData.csv")
metafile <- here("data/metadata.csv")
outdir <- here("results")
```

##Read data
```{r}
# Read metadata and transformed expression data
expr.df <- read.csv(transfcsv, check.names=FALSE)
metadata <- read.csv(metafile)
```

##Marker selection
The following code will print all the available markers in the data. \
The markers of interest for clustering can be copied and paste to the next section.

```{r}
# Check the available markers that can be copied and paste to the next section.
for(column in colnames(expr.df)[-1]){
  cat(paste("\"", column, "\",", sep=""), "\n")
}
```


```{r}

# Select the markers to be used for clustering and dimension reduction.
sel.markers <- c(
#copy selected markers from above here as:
"marker1",
"marker2",
"marker3"
#etc.
)

###################################################################

# Create data list with expression matrix and metadata
data <- list(expr.mat = as.matrix(expr.df[,sel.markers]),
             metadata = data.frame(matrix(nrow = nrow(expr.df), ncol = 0)))

# Add metadata information
for (column in colnames(metadata)) {
  data$metadata <- cbind(data$metadata, metadata[match(expr.df$Sample_ID, metadata[,1]),column])
  colnames(data$metadata)[ncol(data$metadata)] <- column
}

```

##SOM clustering
The following code will perfom SOM clustering to the data considering the previously selected markers. \
The results depend on the indicated number of clusters as well as SOM parameters that can be adjusted.

```{r}

# Adjust nClus (number of clusters) as needed
number_of_clusters <- 8

# Run flowsom on transformed fcs files
flowsom <- FlowSOM(input = data$expr.mat,
                   transform = FALSE,
                   scale = TRUE,
                   colsToUse = NULL, 
                   nClus = number_of_clusters,
                   xdim = 3, #adjust as needed
                   ydim = 3, #adjust as needed
                   rlen = 50, #adjust as needed
                   alpha = c(0.05, 0.01), #adjust as needed
                   seed = seed) 

#########################################################################

# Add the cluster IDs to the data frame
data$metadata$Clusters <- as.factor(flowsom$map$mapping[,1])
levels(data$metadata$Clusters) <- flowsom$metaclustering

```

##tSNE
This code will perform tSNE dimensional reduction for visualization of the clusters. \
Parameters can be adjusted as needed.

```{r}

#tSNE
tsne.obj <- Rtsne(scale(data$expr.mat),
                  dims = 2,
                  initial_dims = ncol(data$expr.mat),
                  perplexity = 50, # Adjust as needed
                  theta = 0.5, # Adjust as needed
                  eta = 200, # Adjust as needed
                  pca = FALSE,
                  normalize = TRUE,
                  max_iter = 1000, # Adjust as needed
                  exaggeration_factor = 12,
                  verbose = FALSE)

###################################################################

# Add tSNE values to the data
tsne.values <- tsne.obj$Y
colnames(tsne.values) <- c("tSNE_1", "tSNE_2")
data$metadata[,colnames(tsne.values)] <- tsne.values
```

##tSNE plot
A 2D plot will be printed. \
The plot can be split by any feature in the metadata.
```{r}
# For splitting the plot:
split.by <- NULL #or any feature from metadata.

# Plot tSNE
tSNEclusters <- plot.feature(data$metadata, 
                              x = "tSNE_1",
                              y = "tSNE_2",
                              color.by = "Clusters", #or any other feature in the metadata
                              split.by = split.by,
                              size = 1, #size of the dots
                              alpha = 1,
                              show.cluser.id = TRUE,
                              show.cluser.id.size = 4,
                              colors = "default", #or a vector of colors with length = nClusters
                              main = "tSNE clusters")

print(tSNEclusters)

```

##UMAP
This code will perform UMAP dimensional reduction for visualization of the clusters. \
Number of neighbours can be adjusted as needed.

```{r}
#UMAP
umap.config <- umap.defaults
umap.config$n_neighbors <- 80 #Adjust as needed
umap.data <- umap(scale(data$expr.mat), config=umap.config)
#Obtain the UMAP values
umap.values <- umap.data$layout
colnames(umap.values) <- c("UMAP_1", "UMAP_2")
#Insert UMAP values to metadata dataframe
data$metadata[,colnames(umap.values)] <- umap.values
```

##UMAP plot
A 2D plot will be printed. \
The plot can be split by any feature in the metadata.
```{r}
# For splitting the plot:
split.by <- NULL #or any feature from metadata.

# Plot UMAP
UMAPclusters <- plot.feature(data$metadata, 
                              x = "UMAP_1",
                              y = "UMAP_2",
                              color.by = "Clusters", 
                              split.by = split.by,
                              size = 1,
                              alpha = 1,
                              show.cluser.id = TRUE,
                              show.cluser.id.size = 4,
                              colors = "default", # or a vector of colors with length = nClusters
                              main = "UMAP clusters")

print(UMAPclusters)
```

##Save output
The output rds file will be saved in the pre-established output directory inside RDSfiles directory.
```{r}

# Create output directory if it does not exist
if (!file.exists(outdir)){
    dir.create(outdir)
}

RDSdir <- file.path(outdir, "RDSfiles")

if (!file.exists(RDSdir)){
    dir.create(RDSdir)
}

AnalysisID <- Sys.Date()
filename <- paste("SOM_clustering_", seed, "_", AnalysisID, ".rds", sep="")
saveRDS(data, file.path(RDSdir, filename))

cat("output file in", RDSdir, "\n")
cat(filename, "\n")
```
