==> analysis/03_Annotation_Visualization.Rmd <==
---
title: "Annotation and Visualization"
author: "Carolina"
date: "2025-01-27"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load packages and utilities

```{r, warning=FALSE}

library(tidyverse)
library(Rtsne)
library(umap)
library(pheatmap)
library(viridis)
library(scales)
library(reshape2)
library(svglite)
library(here)

source(here("src/Plot_functions_flow.R"))

seed <- 1234
set.seed(seed)

```

##Paths to directories and files
*datapath* is the path to the rds file containing the analyzed data and clustering information. \
*outdir* is the path to the output directory that will contain the output results and plots.

```{r}

# Set directories
datapath <- here("results/RDSfiles/Pheno_clustering_1234_2025-01-27.rds")
outdir <- here("results")
```

##Read data
```{r}
# Read data and transformed expression data
data <- readRDS(datapath)

```

##Heatmat
The following code will generate a heatmap showing the expression of each marker by each cluster. \
A gradient color can be specified.

```{r}

# Set gradient colors
myColor <- inferno(100)

# Obtain clusters information
clusterID <- as.factor(data$metadata$Clusters)

#Obtain the expression median for each cluster
dataheat <- data.frame(matrix(0, ncol=ncol(data$expr.mat), nrow=length(levels(clusterID))))
for(i in seq_len(ncol(data$expr.mat))){
  dataheat[,i] <- tapply(data$expr.mat[,i], clusterID, median)
}

colnames(dataheat) <- colnames(data$expr.mat)
dataheat <- data.frame(t(dataheat))
colnames(dataheat) <- levels(clusterID)

# Plot heatmap
heat.expr <- pheatmap(dataheat, 
                      color = myColor, 
                      scale = "row", 
                      border_color = NA, 
                      fontsize = 15, 
                      treeheight_row = 0)  


```

Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.

```{r}
# For splitting the plot:
split.by <- NULL #or any feature from metadata.

# 2D expression plots
expr2D <- lapply(colnames(data$expr.mat), function(x){
                  plot.expr(data$metadata, 
                            data$expr.mat,
                            x = "tSNE_1",
                            y = "tSNE_2",
                            split.by = split.by,
                            color.by=x, 
                            main = x, 
                            size = 1,
                            limits=c(quantile(data$expr.mat[,x], 0.01), 
                                     quantile(data$expr.mat[,x], 0.99))
                  )
                            
})

names(expr2D) <- colnames(data$expr.mat)

for(plot in expr2D) {print(plot)}
```

##Optional: rename clusters

```{r}
# Change cluster labels
# clusters can also be joined in the same population.

new.labels <- c(
  #Rename clusters as:
  "1"= "<population_A>",
  "2"= "<population_B>",
  "3"= "<population_C>"
  #etc
  )

new.order <- c(
  #Reorder clusters as:
  "2",
  "3",
  "1"
  #etc
)

####################################################

data$metadata$Population <- new.labels[as.character(data$metadata$Clusters)]
data$metadata$Population <- factor(data$metadata$Population, levels = new.labels[new.order])

cat("New cluster names are in the column 'Population' of the metadata", "\n")

```

####2D plot
```{r}
#Indicate the features to plot
color.by = "Clusters" #or any feature in metadata
split.by = NULL #or any feature in metadata

# Optional: assign new colors to populations.
new.colors <- c(
  "default"
  # Or specify custom colors for each cluster/population as:
  #"mycolor1", 
  #"mycolor2", 
  #"mycolor3", 
  #etc.
)

#####################################################

annotated2D <- plot.feature(data$metadata, 
                              x = "UMAP_1",
                              y = "UMAP_2",
                              color.by = color.by, 
                              split.by = split.by,
                              size = 1,
                              alpha = 1,
                              show.cluser.id = TRUE,
                              show.cluser.id.size = 2,
                              colors = new.colors, 
                              main = "2D Populations")

print(annotated2D)

```


```{r}
#Indicate the features to plot
x.data <- "Clusters"

# Metadata features to evaluate and respective colors
y.data.colors <- list(SubjectID= "default",
                     Condition= c("#d3caa7", "#2a580e")) 

#################################################################

# Total number of cells per cluster
t.pop <- table(data$metadata[,x.data]) 
print(t.pop)

# Number of cells per cluster distributed in groups
table.list <- lapply(names(y.data.colors), function(x){
  table(data$metadata[,x], 
        data$metadata[,x.data])})
names(table.list) <- names(y.data.colors)

#Barplot of the frequencies of each population in each group
bar.plots <- lapply(names(y.data.colors), function(x){
  bar.plot(table.list[[x]], 
           xlabel=x.data,
           fill.by = x, 
           colors= y.data.colors[[x]])
  }
)

names(bar.plots) <- names(y.data.colors)

for(plot in bar.plots) {print(plot)}


```


##Density plots

```{r}

# Select the factors to plot
subset.by <- "Clusters" #or any other variable in metadata
selected <- c("1", "2") # selected factors to be compared in overlaid density plots
dens.colors <- c("#c2b984", "#953a1c")

#####################################################################

# Subset metadata and expression matrix based on selected categories
sub.data <- lapply(data, function(x){
    subset(x, data$metadata[,subset.by] %in% selected)
    }
  )

# Overlaid density plot
density.plots <- lapply(colnames(sub.data$expr.mat), function(x){
                    dens.plot(sub.data$metadata,
                              sub.data$expr.mat,
                              x = x,
                              color.by = subset.by,
                              linew = 0.5,
                              alpha=0.6,
                              colors = dens.colors,
                              xlabel = x,
                              aspect.ratio = 1.5)
                }
)

names(density.plots) <- colnames(data$expr.mat)

for(plot in density.plots) {print(plot)}

```

##Save output
The output rds file and csv tables will be saved in the pre-established output directory.
```{r}

# Create output directory if it does not exist
if (!file.exists(outdir)){
    dir.create(outdir)
}

RDSdir <- file.path(outdir, "RDSfiles")

if (!file.exists(RDSdir)){
    dir.create(RDSdir)
}

csvdir <- file.path(outdir, "tables")

# Create tables directory if it doesn't exist.
if (!file.exists(csvdir)){
    dir.create(csvdir)
}

# Set analysis ID
AnalysisID <- Sys.Date()

# Save annotated data
saveRDS(data, paste(RDSdir, "/Annotated_", seed, "_", AnalysisID, ".rds", sep=""))

# Save tables as csv files
write.csv(t.pop, file= paste(csvdir, "/Pop-nCells_", seed, "_", AnalysisID, ".csv", sep="")) 

for (i in names(table.list)){
  write.csv(table.list[[i]], 
            file= paste(csvdir,
                        "/Pop-nCells-",
                        i,
                        "_",
                        seed,
                        "_",
                        AnalysisID,
                        ".csv",
                        sep=""))
}
```

##Save plots as svg
The output rds file will be saved in the pre-established output directory inside plots directory.

```{r}

plotdir <- file.path(outdir, "plots")

# Create output directories if they don't exist.
if (!file.exists(plotdir)){
    dir.create(plotdir)
}

# Save plots as svg files
svglite(paste(plotdir, "/Heatmap_", AnalysisID, ".svg", sep=""))
print(heat.expr)
dev.off()

for (i in 1:length(expr2D)){
  svglite(paste(plotdir, "/Expr2D_", names(expr2D)[i], "_", AnalysisID, ".svg", sep = ""))
  print(expr2D[[i]])
  dev.off()
}

svglite(paste(plotdir, "/Annotated2D_", AnalysisID, ".svg", sep = ""))
print(annotated2D)
dev.off()

for (i in 1:length(bar.plots)){
  svglite(paste(plotdir, "/Barplot_", names(bar.plots)[i], "_", AnalysisID, ".svg", sep = ""))
  print(bar.plots[[i]])
  dev.off()
}

for (i in 1:length(density.plots)){
  svglite(paste(plotdir, "/Density_", names(density.plots)[i], "_", AnalysisID, ".svg", sep = ""))
  print(density.plots[[i]])
  dev.off()
}

```
